requirements.txt
fastapi
uvicorn
stripe
python-dotenv
python-jose
sqlalchemy
psycopg2-binary